requirements.txt
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
dask==2023.12.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
grpcio==1.60.0
idna==3.6
importlib-metadata==7.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
locket==1.0.0
modin==0.26.0
msgpack==1.0.7
numpy==1.26.2
packaging==23.2
pandas==2.1.4
partd==1.4.1
protobuf==4.25.1
psutil==5.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.8.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
six==1.16.0
toolz==0.12.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
